+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
-- Fix wrong parse of source-paths for bb.edn when :paths contains symbols not only strings. #507
-- Fix find-definition in jars for cljc files where the var is available on both cljs and clj files. #509
-- Add clojure.core.async common vars to common-refers to be required via code action.
+- General
+  - Fix wrong parse of source-paths for bb.edn when :paths contains symbols not only strings. #507
+
+- Editor
+  - Fix find-definition in jars for cljc files where the var is available on both cljs and clj files. #509
+  - Add clojure.core.async common vars to common-refers to be required via code action.
+  - Remove diagnostics when files are deleted, properly cleaning server. #513
 
 ## 2021.08.05-19.36.09